baselines_src/ALoRa/requirements.txt
matplotlib==3.5.3
numpy==1.21.6
pandas==1.3.5
Pillow==9.5.0
plotly==5.18.0
scikit_learn==1.0.2
scipy==1.7.3
seaborn==0.10.0
tensorflow==2.11.0
torchviz==0.0.3
xlsx2csv==0.8.4
